feat: enhance RAG system with HNSW algorithm support and CLI utilities
Major improvements to the RAG (Retrieval-Augmented Generation) system:

Database & Storage Enhancements:
- Add HNSW (Hierarchical Navigable Small World) algorithm support for efficient vector similarity search
- Implement SQLiteVectorStoreHNSW with optimized HNSW graph structure
- Add new schema definitions for HNSW vector storage with Float32Array support
- Enhance database migration and configuration for HNSW parameters (M, Ml, EfSearch)

CLI Command Utilities:
- Add 'index' command for indexing plugins with configurable embedding models
- Add 'search' command for semantic search across indexed plugins
- Add 'list-collection' command to display all available vector collections
- Support for various embedding models and API configurations

API & Export Improvements:
- Simplify RAG API with cleaner, more intuitive interfaces
- Optimize plugin RAG manager with better metadata handling
- Enhance export/import functionality for vector collections
- Remove deprecated plugin_indexer.go in favor of new command structure

Algorithm & Performance:
- Implement pure Go HNSW algorithm with cosine similarity calculations
- Add distance function abstraction for different similarity metrics
- Optimize vector operations and graph traversal performance
- Add comprehensive test coverage for HNSW operations

Breaking Changes:
- Updated vector store schema to support HNSW-specific fields
- Modified collection configuration structure
- Enhanced embedding interface for better model compatibility

This update significantly improves the RAG system's search performance and provides
a complete CLI toolkit for managing vector collections and plugin indexing.

+++ common/ai/rag/sqlitedb_store_test.go
@@ -24,13 +24,13 @@ func TestSQLiteVectorStore(t *testing.T) {
 			ID:        "doc1",
 			Content:   "Yaklang是一种安全研究编程语言",
 			Metadata:  map[string]any{"source": "Yaklang介绍"},
-			Embedding: []float64{1.0, 0.0, 0.0},
+			Embedding: []float32{1.0, 0.0, 0.0},
 		},
 		{
 			ID:        "doc2",
 			Content:   "RAG是一种结合检索和生成的AI技术",
 			Metadata:  map[string]any{"source": "RAG介绍"},
-			Embedding: []float64{0.0, 1.0, 0.0},
+			Embedding: []float32{0.0, 1.0, 0.0},
 		},
 	}
